enable toast everywhere
