+++ client/src/app/app.tsx
@@ -1,7 +1,8 @@
 import Background from 'components/background'
 import {CurrentDropdown} from 'components/dropdown/dropdown'
 import LostConnection from 'components/lost-connection'
-import {Toaster} from 'components/toast/toast'
+import Toast from 'components/toast'
+import {ToastContainer} from 'components/toast/toast'
 import {CurrentTooltip} from 'components/tooltip/tooltip'
 import {getSettings} from 'logic/local-settings/local-settings-selectors'
 import {localMessages, useMessageDispatch} from 'logic/messages'
@@ -9,6 +10,7 @@ import {
 	getDropdown,
 	getPlayerName,
 	getSession,
+	getToast,
 	getTooltip,
 } from 'logic/session/session-selectors'
 import {getSocketStatus} from 'logic/socket/socket-selectors'
@@ -33,6 +35,7 @@ function App() {
 	const playerName = useSelector(getPlayerName)
 	const socketStatus = useSelector(getSocketStatus)
 	const connected = useSelector(getSession).connected
+	const toastMessage = useSelector(getToast)
 	const tooltip = useSelector(getTooltip)
 	const dropdown = useSelector(getDropdown)
 	const settings = useSelector(getSettings)
@@ -42,7 +45,6 @@ function App() {
 	const [menuSection, setMenuSection] = useState<string>(
 		lastMenuSection || 'main-menu',
 	)
-	let enableToast = false
 
 	const menuSectionSet = (section: string) => {
 		setMenuSection(section)
@@ -134,7 +136,19 @@ function App() {
 					tooltipWidth={tooltip.tooltipWidth}
 				/>
 			)}
-			{enableToast && <Toaster />}
+			<ToastContainer>
+				{toastMessage.map((toast, i) => {
+					return (
+						<Toast
+							title={toast.toast.title}
+							description={toast.toast.description}
+							image={toast.toast.image}
+							id={toast.id}
+							key={i}
+						/>
+					)
+				})}
+			</ToastContainer>
 			{playerName && !socketStatus && <LostConnection />}
 		</main>
 	)
